Implement "input filters"
These are commands and other goodies. Essentially you register a
function that takes a text argument, and you can filter whether you
should let it bubble all the way through to "send". Eventually it should
allow editing the text, but for now it doesnt. Should be as simple as
changing the apply-partially, just passing the string in to the
initial-value and short-circuiting if a function returns nil...