Auto-load the insert-read-only macro
It feels like there may be a race condition between the matrix-client
running and the load being resolved? Idk, it's weird, and I'm going
to try to ignore it and see if autoloading this function will help. At
the very least, if there is a race-condition where mclient-handlers are
being run before matrix-client is loading, this will trigger the module
to load.