Fix last-ts calculation when slot is unbound
Previously I added a check for last-ts to ensure that if a message
hadn't been received it doesn't throw a type error. The issue with that
is that if no messages were ever received `last-ts` would be nil, so the
watchdog would be cancelled without trying to reconnect.

This changes the last-ts to just be "0" if it isn't set, so it will
always try to reconnect if no messages have been received.

There might be a better way to do this, but it needs to be fixed as it's
currently broken (sorry!)