Change: Rewrite insertion and prompt
Every string inserted into a room buffer now has a timestamp
text-property.  This is used to find the correct place when inserting
messages into a buffer.  This will allow us to later add code that
fetches earlier messages and inserts them at the right place, as well
as, perhaps, more async code here and there, because events no longer have to
be received in-order to be inserted at the right place.

The prompt has also been rewritten to use an overlay and have a
customizable string.  The overlay is used to find the prompt, so the
string doesn't matter.