Change: (matrix-client-update-name) Prioritize room alias over 1-1
If a room has a public alias but only has 2 members, the alias should be
used rather than the other member's name, because it's a public room and
we expect other people may join.  e.g. #emacs-matrix-client:matrix.org
only has me and @jgkamat at the moment, but we hope it will have more
someday!  :)