Add: Pretend buffers are file-backed
We want room buffers to be marked as modified when they have unseen
messages so that packages like Helm can distinguish them.  However, Helm
also needs to distinguish them from other modified, non-file-backed
buffers, such as *scratch*, because they are always considered modified,
and Helm does not want to highlight them as modified buffers, since they
are always considered modified.

To fix that, we set buffer-file-name for our room buffers to an empty
string.  This way Helm shows it as modified, like a modified,
file-backed buffer, but Helm can still show *scratch* normally.

However, if buffer-file-name is set to the empty string for a buffer,
the rename-buffer function ALWAYS uses the name "-", which is not what
we want.  So we let-bind buffer-file-name to nil around the call to
rename-buffer, causing it to use the name we want.