Change: Canonicalize user ID on login and use full IDs
The previous commit allowed for storing the user ID and token, which
stores the full, canonicalized user ID, and then loads it into the
connection object as :username.  That broke code which refers to the
user's own name to determine whether an event is from the local user or
others.  This commit updates that code to use the full, canonicalized
user ID instead of the bare username (e.g. "@alphapapa:matrix.org"
instead of "alphapapa").