Add/Fix: Persist transaction ID and initialize to random number
NOTE: According to
<https://matrix.org/docs/spec/client_server/r0.3.0.html#id183>, the
transaction ID should be scoped to the access token, so we should
preserve it with the token.  However, if the client crashes and fails
to save the TID, and then reuses it in the future...what happens?  The
server seems to accept messages with already-used TIDs.  Maybe it has
some kind of heuristic...  I found these:
<https://github.com/matrix-org/synapse/issues/1481> and
<https://github.com/matrix-org/synapse/pull/1624>.