Add: Display sending messages immediately, with button to resend
WIP

WIP

WIP: Progress

Now, when sending a message, it is immediately displayed in the
buffer, with normal faces and metadata, along with a "[pending]"
button.  When the server responds with the event ID, it's added to the
message, and the button is removed.  Then when the event is received
back from the server in the next /sync, the event is deleted from the buffer and
replaced with the event from the server, which has the server's
origin timestamp.

Still need to test resending messages that don't get sent, but now
that the matrix.org homeserver is a bit more reliable, with the
improvements they've recently made, this doesn't happen as often...