Fix: (matrix-url-with-retrieve-async) Error when connection completely fails
It seems that the callback can be called with `nil' when the
connection fails before getting any headers, like:
url-http-end-of-document-sentinel(#<process matrix.org<5>> "connection
broken by remote peer\n"), in which case `url-http-end-of-headers' is
nil, so we need to call the error fn.  Would like to structure this
more cleanly.