Fix: UTF-8 encode JSON data
Fixes #50.  See also <https://github.com/magit/ghub/issues/12>,
<https://github.com/dakrone/es-mode/issues/62>, and
<https://github.com/magit/ghub/issues/35>.

I can't explain why this is necessary.  We encode the message to UTF-8
already, and there shouldn't be anything multi-byte about the JSON
syntax.  And if there were, shouldn't that have been the case already,
not just when sending multi-byte content encoded in the JSON?
e.g. see especially
<https://github.com/magit/ghub/issues/35#issuecomment-348832926>.
Jonas's comment on commit
<https://github.com/magit/ghub/commit/45cd8c8341f2065aad6ed72200c10842f8087498>
almost seems to explain it, but I still don't understand why it would
happen here, because we were already encoding the messages to UTF-8
before giving them to url.el.

Anyway, this seems to fix it.

And this is another thing that I think should be fixed in the
url-with-retrieve-async macro.  We should just encode EVERYTHING to
UTF-8 and tell the server that we're using it.  If that were already
the case in url.el, none of these bugs would have happened, and it
would have saved so many man-hours in all these projects.