Change: Use request instead of url.el
It's impossible to upload files with url.el unless
url-http-create-request is overridden with a fixed version, and that
must be done with function advice (which applies to everything in
Emacs), because doing it with cl-letf doesn't work (who knows why--it
works fine for other things).  Obviously we don't want to override
that function globally, even though it works fine now, because when
url.el changes in the future, that might break something.  And it's
not a good idea, anyway.

So let's switch back to request.el.  Frankly, I don't even remember
why I switched to my url-with-retrieve-async function.  There is this
pesky bug at https://github.com/tkf/emacs-request/issues/92, but that
only applies to synchronous requests, which we aren't using.