Revert: "Change: Use request instead of url.el"
This reverts commit 21c8af26d899f29885060b3ef52f644a86531a41.

request.el has a bug which causes our sync requests to get duplicated.
The last message I received in a room arrived 21 times because of 21
simultaneous sync requests.  I don't know what is causing it, and
request.el is very difficult to debug.  So we will have to go back to
using url.el for everything we can use it for.  For a few things that
don't work with url.el, like uploading files, we'll have to use
request.el.

What a mess.  And how disappointing that there is no HTTP library for
Emacs that doesn't have these kinds of virtually-impossible-to-debug
bugs.

See https://github.com/jgkamat/matrix-client-el/pull/27#issuecomment-428689369