Change: Improve buffer naming, especially on initial sync
In large rooms (e.g. #emacs with > 600 users), initial sync is now
faster because 1) the room name is only computed twice (when the room
buffer is created, and when the initial sync is completed) rather than
for every user "join" event (and the initial computation is very fast,
because there are one or no users known to be in the room when the
buffer is first created), and 2) we short-circuit the computations
when possible, avoiding looking up displaynames for users
unnecessarily.