Change: Improve insertion of headers for old messages
Probably need to refactor a bit, or add some functionality to
ordered-buffer to cleanly handle inserting from both ends.  But this
isn't bad.  The only issue now is that, when inserting old messages, a
new timestamp header is always inserted, even if the timestamp
difference is within the delta setting, which is not necessarily a bad
thing, because it makes it easier to see where the newly fetched
messages are.

It gets a little complicated because, when inserting old messages, if
we insert from the top (which makes sense), we need to insert the
messages in a different order so the prefix-fn looks at the right
event to determine whether to insert a new header.