Add: Drag-and-drop support
This allows URLs to be drag-and-dropped onto room buffer windows, upon
which it will be offered to upload them to the room (e.g. for an
image).  This works from e.g. Dolphin, Firefox, etc.

However, it does not work from Google Chrome/Chromium due to a nearly
4-year-old bug that Google has neglected to fix, and has since closed
as "old, likely obsolete."  See
<https://bugs.chromium.org/p/chromium/issues/detail?id=461390>,
<https://debbugs.gnu.org/cgi/bugreport.cgi?bug=19885>.  The latter
contains a workaround users may apply to their Emacs config.