Change: Simplify room avatar callback
It's not necessary to use insert-image in a temp buffer and get the
buffer as a string.  Just propertize a string with the image as the
display property.

Also, use a non-breaking space instead of a regular space to hold the
image.  That way, when used as a prefix for buffer names (see next
commit), it won't cause Emacs to hide the buffer by default.