Fix: apply-if-fn macro
When the files are byte-compiled, the use of apply-if-fn in
matrix-client-update in matrix-client-room.el leaves the fn-name
variable undefined.  This only happens when the files are
byte-compiled, e.g. by package.el or quelpa.  Maybe it happens because
the files are compiled in a different order than I would expect.
Moving the macro into matrix-macros.el seems to fix the problem.  And
I required that file in matrix-client-room.el just to be safe, even
though it seems to work without doing that.