Change: Rewrite input-sending, quoting, replying
With Org syntax enabled by default, quoting would include more than it
should, including syntax, etc.

This rewrites almost all sending, quoting, and replying code.  It more
closely imitates Riot, but with one difference: when quoting part of a
message, we still send the event ID of the quoted message, because
that seems useful.  It does cause Riot to display the entire quoted
message above the quoted part, but that seems like something Riot
should fix; what's the point of being able to send the IDs of quoted
messages if you don't actually do it when "quoting", but only when
"replying"?