Change: All methods to defuns
Chris Wellons (hi @skeeto! ;) has reminded me that method dispatch in
Emacs is slow, and in this case we gain nothing by using methods
anyway, so we should use defuns instead.  Maybe this will help a
little bit with initial sync performance.

Emacs makes this so easy.  One occur (or in this case, rg) buffer
showing "cl-defmethod" matches, one keyboard macro that selects an
occurrence, replaces and raises a few sexps with lispy, then switches
back to the rg buffer and goes to the next item--and then all I have
to do is press F4 repeatedly until it's done.