Add/Change: Watchdog timer
And associated changes.  This should make syncing much more reliable.
It seems to be working well.

I wish this weren't necessary, but it seems to be, to work around bugs
somewhere in the stack that prevent some network processes from timing
out and calling the error callback.

Thanks to @damiencassou for his help testing.