WIP: magit-section sidebar group
This is a lot of gain for very little change.  With a bit more
polishing and tweaking, it could be great.

However, I'm reluctant to base this on magit-section since this isn't
a Magit-related package, and magit-section isn't intended to be used
by other packages.  magit-section could be copied into this project
and renamed, and that would probably work fine, but then it would
diverge from magit-section, which might be undesirable.  magit-section
is unlikely to be factored out of Magit anytime soon, so this should
be carefully considered.  The built-in Emacs widget library might be
an alternative, although it looks more complicated.